app/requirements.txt
flask
matlabengine==9.13.1